stm32/usart: unify v1 and v2
